Created Terms of Use page for user registration step and empty contents of Terms and conditions page in preparation of terms and conditions from the clients.
TODO: Get info of which pages the terms of use page will go to if back button is clicked. Also, work with Vlad to put in the Terms of Use page for user registration.

The terms of Use files will be deleted because i will pass them to Vlad and his team to plug it in to make the merge easier.